==> 04_finalJoin.ipynb <==
%run ./01_acquireList.ipynb


# Minimal run in Parish mode:
results = run(
    parish="All",               
    start_date="09/01/2025",            
    end_date="09/02/2025",              
    keywords=[""],  # optional
    headless=False # set True for headless
)

df = pd.DataFrame(results)
listDF = df.drop(columns=["status", "validated_date", "decision", "detail_link"], errors="ignore")
listDF = listDF.rename(columns={
    "proposal": "address",
    "address": "description"

})

app_ids = listDF["application_number"].astype(str).tolist()

%run ./02_acquireFileAndDocURL.ipynb


df = files_df

%run ./03_processDocumentContent.ipynb


finalDF = pd.merge(listDF, contentDF, on="application_number", how="left")


from datetime import datetime
date_str = datetime.now().strftime("%d%B%Y")
filename = f"final{date_str}.csv"
finalDF.to_csv(filename, index=False)
